printing even numbers

+++ 0x05-pointers_arrays_strings/6-puts2.c
@@ -8,7 +8,7 @@
 
 void puts2(char *s)
 {
-	int str, i, j = 0;
+	long int str, i, j = 0;
 	char *x = s;
 
 	while (*x != '\0')
